refactor: rename messenger classes
