+++ packages/siwx/src/index.ts
@@ -1,2 +1,2 @@
-export * from './messages/index.js'
+export * from './messengers/index.js'
 export * from './verifiers/index.js'
