--- packages/siwx/src/messages/index.ts
@@ -1,1 +1,0 @@
-export * from './InformalMessage.js'
