+++ packages/siwx/src/messengers/InformalMessenger.ts
@@ -1,5 +1,5 @@
 import type { SIWXMessage, SIWXMessageMethods } from '@reown/appkit-core'
-import { SIWXMessenger } from '../core/SIWXMessager.js'
+import { SIWXMessenger } from '../core/SIWXMessenger.js'
 
 /**
  * Follow Informal Message template from EIP-4361
@@ -27,7 +27,7 @@ import { SIWXMessenger } from '../core/SIWXMessager.js'
  * - ${resources[n]}
  * ```
  */
-export class InformalMessage extends SIWXMessenger {
+export class InformalMessenger extends SIWXMessenger {
   protected readonly version = '1'
 
   protected override stringify(params: Omit<SIWXMessage, keyof SIWXMessageMethods>): string {
