+++ packages/siwx/src/messengers/index.ts
@@ -1,0 +1,1 @@
+export * from './InformalMessenger.js'
